+++ plugin/speedb/pinning_policy/scoped_pinning_policy.cc
@@ -16,8 +16,12 @@
 
 #include "plugin/speedb/pinning_policy/scoped_pinning_policy.h"
 
+#include <inttypes.h>
+
+#include <cstdio>
 #include <unordered_map>
 
+#include "port/port.h"
 #include "rocksdb/utilities/options_type.h"
 
 namespace ROCKSDB_NAMESPACE {
@@ -66,6 +70,28 @@ bool ScopedPinningPolicy::CheckPin(const TablePinningOptions& tpo,
   }
 
   return true;
+}
+
+std::string ScopedPinningPolicy::GetPrintableOptions() const {
+  std::string ret;
+  const int kBufferSize = 200;
+  char buffer[kBufferSize];
+
+  snprintf(buffer, kBufferSize,
+           "    pinning_policy.capacity: %" ROCKSDB_PRIszt "\n",
+           options_.capacity);
+  ret.append(buffer);
+
+  snprintf(buffer, kBufferSize,
+           "    pinning_policy.last_level_with_data_percent: %" PRIu32 "\n",
+           options_.bottom_percent);
+  ret.append(buffer);
+
+  snprintf(buffer, kBufferSize, "    pinning_policy.mid_percent: %" PRIu32 "\n",
+           options_.mid_percent);
+  ret.append(buffer);
+
+  return ret;
 }
 
 }  // namespace ROCKSDB_NAMESPACE